frescomAPIs-main/frescomAPIs-main/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
cffi==1.17.0
click==8.1.7
cryptography==43.0.0
ecdsa==0.19.0
fastapi==0.112.0
fastapi-pagination==0.12.26
h11==0.14.0
idna==3.7
oracledb==2.3.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.10
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.5
